requirements.txt
Django==4.2.8
djangorestframework==3.14.0
drf_yasg==1.21.7
python-dotenv==1.0.0
django-extensions==3.2.3
psycopg2==2.9.9
gunicorn==21.2.0
